requirements_backend.txt
tensorflow==2.15.0
fastapi==0.109.0
numpy>=1.23.5,<2.0.0
pillow==10.0.1
opencv-python-headless
matplotlib==3.8.0
scipy>=1.8,<1.9.2
scikit-image==0.20.0
scikit-learn==1.2.2
ipywidgets==8.0.3
uvicorn==0.26.0
python-multipart









